Tie update menu into StatueTower
This is implemented in a somewhat special way. Upgrade details
for a tower are stored in a nested Dictionary structure inside
a variable on the tower caled "upgrades_table" this is then read
by the menu to create a set of cusomized CheckButton nodes inside
the menu.

That all essently just means it should be easy to integrate the menu
into the cannon tower!